kernel/web/pyproject.toml
[project]
version = "1.0.0"
name = "leap360-web"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "sqlmodel>=0.0.24",
    "cryptography>=45.0.6",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.116.1",
]
